+++ nlp/docs/base_models.md
@@ -1,0 +1,9 @@
+# Base Models
+
+This module contains the base models used in the application.
+
+```{eval-rst}
+.. automodule:: src.base_models
+   :members:
+   :undoc-members:
+   :show-inheritance:
