+++ nlp/docs/index.md
@@ -18,6 +18,9 @@ main.md
 google_service.md
 version_checker.md
 authorization.md
+base_models.md
+models_loading.md
+utils.md
 ```
     
 ## Indices and tables
