+++ nlp/docs/models_loading.md
@@ -1,0 +1,9 @@
+# Model loading
+
+This module contains the loading functions for the models used in the application.
+
+```{eval-rst}
+.. automodule:: src.models_loading
+   :members:
+   :undoc-members:
+   :show-inheritance:
