+++ nlp/docs/utils.md
@@ -1,0 +1,9 @@
+# Utilities
+
+This module contains utility functions used in the application.
+
+```{eval-rst}
+.. automodule:: src.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
