+++ nlp/src/models_loading.py
@@ -1,0 +1,54 @@
+import os
+import fasttext
+from transformers import AutoTokenizer, AutoModelForSequenceClassification, pipeline
+from keybert import KeyBERT
+import src.globals as g
+
+
+
+def load_model(model_name: str):
+    """
+    Load spam models function. It loads the model and tokenizer for the given
+    model name.
+
+    Args:
+         model_name (str): Name of the model to load.
+    """
+    model_path = os.path.join("models", model_name, "v1")
+
+    if not os.path.exists(model_path) and model_name != "keywords":
+        raise RuntimeError("Model file not found")
+
+    if model_name != "keywords" and model_name != "language":
+        try:
+            tokenizer = AutoTokenizer.from_pretrained(model_path)
+            model = AutoModelForSequenceClassification.from_pretrained(
+                model_path)
+        except ValueError as e:
+            raise RuntimeError(f"Failed to load model: {e}")
+    elif model_name == "keywords":
+        g.keyword_model = KeyBERT()
+    elif model_name == "language":
+        try:
+            g.language_model = fasttext.load_model(model_path + "/model.bin")
+        except ValueError as e:
+            raise RuntimeError(f"Failed to load model: {e}")
+
+    if model_name == "spam":
+        g.spam_pipeline = pipeline("text-classification", model=model,
+                                 tokenizer=tokenizer)
+    elif model_name == "sarcasm":
+        g.sarcastic_pipeline = pipeline("text-classification", model=model,
+                                      tokenizer=tokenizer)
+    elif model_name == "sentiment":
+        g.sentiment_model = model
+        g.sentiment_tokenizer = tokenizer
+    elif model_name == "political":
+        g.political_pipeline = pipeline("text-classification", model=model,
+                                      tokenizer=tokenizer)
+    elif model_name == "hate_speech":
+        g.hate_speech_pipeline = pipeline("text-classification", model=model,
+                                        tokenizer=tokenizer)
+    elif model_name == "clickbait":
+        g.clickbait_pipeline = pipeline("text-classification", model=model,
+                                      tokenizer=tokenizer)
